+++ packages/elements-utils/src/toc/toc.ts
@@ -36,14 +36,43 @@ export function generateToC(searchResults: NodeData[]) {
   )();
 }
 
-function traverseToC(items: TableOfContentItem[], apply: (item: Item) => TableOfContentItem[]) {
+export function generateTocSkeleton(searchResults: NodeData[]) {
+  return pipe(
+    () => searchResults,
+    groupNodesByType,
+    ({ articles, models, httpServices }) => {
+      const toc: ITableOfContents = { items: [] };
+
+      // Articles
+      pipe(() => articles, sortArticlesByTypeAndPath, appendArticlesToToC(toc))();
+
+      // HTTP Services
+      pipe(() => httpServices, sortNodesByUri, appendHttpServicesItemsToToC(toc))();
+
+      // Models
+      pipe(() => models, sortNodesByUri, appendModelsToToc(toc))();
+
+      return toc;
+    },
+  )();
+}
+
+function modifyEach(
+  items: TableOfContentItem[],
+  apply: (item: Item) => TableOfContentItem[],
+  shouldReplace?: (item: Item) => boolean,
+) {
   for (let i = items.length - 1; i > -1; i--) {
     const item = items[i];
     if (isItem(item)) {
-      items.splice(i + 1, 0, ...apply(item));
+      if (shouldReplace?.(item)) {
+        items.splice(i, 1, ...apply(item));
+      } else {
+        items.splice(i + 1, 0, ...apply(item));
+      }
     }
     if (isGroup(item)) {
-      traverseToC(item.items, apply);
+      modifyEach(item.items, apply);
     }
   }
 }
@@ -53,8 +82,8 @@ export function injectHttpOperationsAndModels(searchResults: NodeData[], toc: IT
     () => searchResults,
     groupNodesByType,
     ({ models, httpServices, httpOperations }) => {
-      traverseToC(toc.items, ({ uri }) => {
-        const httpService = httpServices.find(httpService => httpService.uri === uri);
+      modifyEach(toc.items, ({ uri }) => {
+        const httpService = httpServices.find(matchesUri(uri));
 
         if (!httpService) return [];
 
@@ -74,6 +103,34 @@ export function injectHttpOperationsAndModels(searchResults: NodeData[], toc: IT
       });
     },
   )();
+}
+
+export function resolveHttpServices(searchResults: NodeData[], toc: ITableOfContents) {
+  pipe(
+    () => searchResults,
+    groupNodesByType,
+    ({ httpServices }) => {
+      modifyEach(
+        toc.items,
+        item => {
+          const httpService = httpServices.find(matchesUri(item.uri));
+
+          if (!httpService) return [];
+
+          return [
+            { type: 'divider', title: item.title },
+            { type: 'item', title: 'Overview', uri: httpService.uri },
+          ];
+        },
+        item => httpServices.some(matchesUri(item.uri)),
+      );
+    },
+  )();
+  injectHttpOperationsAndModels(searchResults, toc);
+}
+
+function matchesUri(uri: string) {
+  return (item: NodeData) => item.uri.replace(/^\//, '') === uri.replace(/^\//, '');
 }
 
 export function groupNodesByType(searchResults: NodeData[]) {
@@ -278,6 +335,16 @@ function appendHttpServiceItemsToToCWithTags(toc: ITableOfContents) {
   };
 }
 
+function appendHttpServicesItemsToToC(toc: ITableOfContents) {
+  return (httpServices: NodeData[]) => {
+    if (httpServices.length) {
+      httpServices.forEach(httpService =>
+        toc.items.push({ type: 'item', title: httpService.name, uri: httpService.uri }),
+      );
+    }
+  };
+}
+
 export function appendHttpServicesToToC(toc: ITableOfContents) {
   return ({
     httpServices,
@@ -322,6 +389,6 @@ export function appendModelsToToc(toc: ITableOfContents) {
 }
 
 function getDirsFromUri(uri: string) {
-  const strippedUri = uri.replace(/^\/(?:docs\/)?/, '');
+  const strippedUri = uri.replace(/^\/?(?:docs\/)?/, '');
   return strippedUri.split(sep).slice(0, -1);
 }