+++ packages/elements/src/components/MarkdownViewer/index.tsx
@@ -1,7 +1,7 @@
 import { IMarkdownViewerProps, MarkdownViewer as MarkdownViewerComponent } from '@stoplight/markdown-viewer';
 import * as React from 'react';
 
-import { useComponents } from '../../hooks/useComponents';
+import { useComponents } from '../../context/Components';
 
 /**
  * Wraps @stoplight/markdown-viewer and passes in custom components from the context provider